logging: Central collection of all the system logs
This patch will add grafana-alloy as logging client
which will be running as systemd service in ghaf-host
and other vms. In admin-vm there will be loki systemd
service running which be collecting and storing logs
in persistent location.
Logs collected can be viewed locally on the target
system or user can push to remote server to view it.

Signed-off-by: Vunny Sodhi <vunny.sodhi@unikie.com>
Signed-off-by: Ville Ilvonen <ville.ilvonen@unikie.com>
Signed-off-by: Vunny Sodhi <vunny.sodhi@unikie.com>